Corrected version main.cpp
The ParseQuery() method returns a Query structure. The exception throwing function has been moved to the ParseQueryWords() function. The exception check has been removed in the MakeUniqueNonEmptyStrings() function and in the SearchServer constructor, which takes a string as an argument; exception checking is carried out by the constructor, where the container is the argument.